Consultation

+++ app/Http/Controllers/API/ManageUserController.php
@@ -515,7 +515,7 @@ class ManageUserController extends BaseController
         // Ambil user dengan role Psikolog ('P') dan muat relasi
         $user = User::where('id', $id)
             ->where('role', 'P')
-            ->with(['psikolog.psikolog_topic.topic', 'psikolog.psikolog_category'])
+            ->with(['psikolog.psikolog_topic.topic', 'psikolog.psikolog_category', 'psikolog.bank'])
             ->whereHas('psikolog', function ($query) {
                 $query->where('is_active', true); // Hanya ambil psikolog yang aktif
             })
@@ -532,6 +532,7 @@ class ManageUserController extends BaseController
 
         // Format detail user
         $psikologDetails = $user->psikolog;
+        // dd($psikologDetails);
         $formattedUser = [
             'id' => $user->id,
             'name' => $user->name,
@@ -542,6 +543,9 @@ class ManageUserController extends BaseController
             'gender' => $user->gender,
             'sipp' => $psikologDetails->sipp ?? null,
             'practice_start_date' => Carbon::parse($psikologDetails->practice_start_date)->translatedFormat('Y-m-d'), 
+            'bank_id' => $psikologDetails->bank->id ?? null,
+            'bank_name' => $psikologDetails->bank->name ?? null,
+            'rekening' => $psikologDetails->account_number ?? null,
             'description' => $psikologDetails->description,
             'selected_topics' => $psikologDetails->psikolog_topic->map(function ($topicRelation) {
                 return [
@@ -597,12 +601,22 @@ class ManageUserController extends BaseController
             'photo_profile' => 'sometimes|image|mimes:jpeg,png,jpg|max:2048',
             'date_birth' => 'sometimes|date',
             'gender' => 'sometimes|in:F,M',
-            'sipp' => 'nullable|string|max:50',
+            'sipp' => 'nullable|string',
             'practice_start_date' => 'nullable|date',
             'updated_topics' => 'nullable|array', 
+            'bank_id' => 'nullable|exists:payment_methods,id',
+            'rekening' => 'nullable|string',
         ], [
-
-            
+            'updated_topics.*' => 'ID topik tidak valid',
+            'email.unique' => 'Email sudah digunakan oleh pengguna lain',
+            'photo_profile.image' => 'Format gambar tidak sesuai.',
+            'photo_profile.mimes' => 'Format gambar harus JPEG, PNG, atau JPG.',
+            'photo_profile.max' => 'Ukuran gambar tidak boleh lebih dari 2 MB.',
+            'date_birth.date' => 'Format tanggal lahir tidak valid.',
+            'gender.in' => 'Jenis kelamin tidak valid.',
+            'sipp.string' => 'SIPP harus berupa teks.',
+            'practice_start_date.date' => 'Format tanggal mulai praktik tidak valid.',
+            'updated_topics.*.exists' => 'ID topik tidak valid.',
         ]);
 
         if ($validator->fails()) {
@@ -637,12 +651,19 @@ class ManageUserController extends BaseController
                 'phone_number' => $validatedData['phone_number'] ?? $user->phone_number,
                 'date_birth' => $validatedData['date_birth'] ?? $user->date_birth,
                 'gender' => $validatedData['gender'] ?? $user->gender,
-                'practice_start_date' => $validatedData['practice_start_date'] ?? $user->practice_start_date,
             ]);
 
             // Update data psikolog
             $psikolog = $user->psikolog;
             if ($psikolog) {
+
+                //Perbarui data psikolog
+                $psikolog->update([
+                    'practice_start_date' => $validatedData['practice_start_date'] ?? $user->practice_start_date,
+                    'bank_id' => $validatedData['bank_id'] ?? $user->bank_id,
+                    'account_number' => $validatedData['rekening'] ?? $user->account_number,
+                ]);
+
                 // Perbarui SIPP dan PsikologPrice
                 if (isset($validatedData['sipp'])) {
                     $sipp = $validatedData['sipp'];
@@ -669,22 +690,24 @@ class ManageUserController extends BaseController
                     );
                 }
 
-               // Cek apakah ada perubahan dalam topik
-               $existingTopicIds = $user->psikolog->psikolog_topic->pluck('topic_id')->toArray();
-               $newTopicIds = $request->updated_topics;
+                // Cek apakah ada data updated_topics pada request
+                if ($request->has('updated_topics')) {
+                    $existingTopicIds = $user->psikolog->psikolog_topic->pluck('topic_id')->toArray();
+                    $newTopicIds = $request->updated_topics;
 
-               // Jika topik berubah, lakukan pembaruan
-               if (array_diff($existingTopicIds, $newTopicIds) || array_diff($newTopicIds, $existingTopicIds)) {
-                   // Hapus topik lama
-                   $user->psikolog->psikolog_topic()->delete();
+                    // Jika topik berubah, lakukan pembaruan
+                    if (array_diff($existingTopicIds, $newTopicIds) || array_diff($newTopicIds, $existingTopicIds)) {
+                        // Hapus topik lama
+                        $user->psikolog->psikolog_topic()->delete();
 
-                   // Tambahkan topik baru
-                   $newTopics = collect($newTopicIds)->map(function ($topicId) {
-                       return ['topic_id' => $topicId];
-                   });
-                   $user->psikolog->psikolog_topic()->createMany($newTopics->toArray());
-               }
-               $user->load('psikolog.psikolog_topic');
+                        // Tambahkan topik baru
+                        $newTopics = collect($newTopicIds)->map(function ($topicId) {
+                            return ['topic_id' => $topicId];
+                        });
+                        $user->psikolog->psikolog_topic()->createMany($newTopics->toArray());
+                    }
+                    $user->load('psikolog.psikolog_topic');
+                }
             }
 
             DB::commit();
@@ -728,8 +751,4 @@ class ManageUserController extends BaseController
             return $this->sendError('Psikolog tidak dapat dihapus, karena dipakai pada tabel lain.', [$e->getMessage()], 500);
         }
     }
-
-
-
-
 }
